// app/model/dataPribadi.ts
export interface DataPribadi {
    perkawinan: string; // Status perkawinan
    provinces: string; // Provinsi
    daerah: string; // Kabupaten/Kota
    email: string; // Alamat email
    telepon: string; // Nomor telepon
    studi: string; // Status studi
    kerja: string; // Status kerja
    pekerjaan?: string; // Pekerjaan (opsional)
}